requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mcp>=1.0.0
fastmcp>=0.1.0
qdrant-client>=1.12.0
numpy>=1.26.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0

# Testing dependencies
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0

# Code quality
black>=23.0
ruff>=0.1.0
mypy>=1.0

# Cache
redis[hiredis]>=5.0.0

# Monitoring & Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-exporter-prometheus>=1.12.0rc1
prometheus-client>=0.19.0
structlog>=23.2.0
python-json-logger>=2.0.7
